refactor App.tsx; remove parsing logic
