+++ src/components/App.tsx
@@ -1,18 +1,13 @@
 import * as React from "react";
 import ClusterView from './views/ClusterView';
 import NodeView from "./views/NodeView";
-import kDeployment from "../kObjects/kDeployment";
-import Container, {env} from "../kObjects/container";
-import kService from "../kObjects/kService";
 import {kObject} from "../kObjects/kObject"
-import statefulContainer, {volumeMount} from "../kObjects/statefulContainer";
-import { kStatefulSet } from "../kObjects/kStatefulSet";
-import volumeClaimTemplates from "../kObjects/volumeClaimTemplates";
-import { createEmitAndSemanticDiagnosticsBuilderProgram } from "typescript";
+import * as kObjects from "../kObjects/__index";
+import * as dataParser from "../component_data/kDataParser";
 
 function App() {
-  const kObjArray: kObject[] = [];
-  const [dataProp, SetDataProp] = React.useState<typeof kObjArray | undefined>([]);
+  let kObjArray: kObject[] = [];
+  const [dataProp, SetDataProp] = React.useState<kObject[]>([]);
   const [nodeViewPage, setNodeViewPage] = React.useState(false);
   const [view, setView] = React.useState('Cluster View')
   const [masterNode, setMasterNode] = React.useState("Kubernetes Cluster")
@@ -26,8 +21,8 @@ function App() {
   //fetch data from backend, push to kDeployArray
   function getData(): void {
     fetch('http://localhost:3000/getData')
-      .then((data: any) => data.json())
-      .then((data: any) => {
+      .then((data: Response) => data.json())
+      .then((data: any[]) => {
         // Data will be an array of objects. Each object represents a different YAML file.
         parseData(data);
       })
@@ -36,9 +31,9 @@ function App() {
 
   function fetchLiveData(): void {
     fetch('http://localhost:3000/statusConditions')
-      .then((data: any) => data.json())
-      .then((data: any) => {
-        data.forEach((data: any) => {
+      .then((data: Response) => data.json())
+      .then((data: string[]) => {
+        data.forEach((data: string) => {
           deploymentStatus.push(data);
         })
         setDeployment(deploymentStatus)
@@ -46,79 +41,10 @@ function App() {
     .catch((error) => console.log('GET /getLiveData response error: ', error));
   }
 
-  function parseData(relevantData: any[]) 
+  function parseData(relevantData: kObject[]) 
   {
-      for(let i = 0; i < relevantData.length; i++){
-      // Since each YAML file can have multiple objects, [0] assumes that there is only one object per file.
-      let ele = relevantData[i][0];
-      // Checks to see if kubernetes object is a deployment 
-      if(ele.kind === 'Deployment') {
-        const newEnv = new env(
-          ele.spec.template.spec.containers[0].env[0].name,
-          ele.spec.template.spec.containers[0].env[0].value,
-        );
-        const newContainer = new Container(
-          ele.spec.template.spec.containers[0].name,
-          ele.spec.template.spec.containers[0].image,
-          newEnv,
-          ele.spec.template.spec.containers[0].ports[0].containerPort,
-        );
-        const newDeployment = new kDeployment(
-          ele.metadata.namespace,
-          ele.kind,
-          ele.metadata.name,
-          ele.spec.template.metadata.labels.name,
-          ele.spec.replicas,
-          newContainer,
-        );
-        kObjArray.push(newDeployment);
-      }
-      else if(ele.kind === 'StatefulSet') {
-        const newVolumeMount = new volumeMount(
-          ele.spec.template.spec.containers[0].volumeMounts[0].mountPath,
-          ele.spec.template.spec.containers[0].volumeMounts[0].name,
-        )
-        const newStatefulContainer = new statefulContainer(
-          ele.spec.template.spec.containers[0].name,
-          ele.spec.template.spec.containers[0].image,
-          ele.spec.template.spec.containers[0].ports[0].containerPort,
-          newVolumeMount
-        )
-        const newVolumeClaimTemplates = new volumeClaimTemplates(
-          ele.spec.volumeClaimTemplates[0].metadata.name,
-          ele.spec.volumeClaimTemplates[0].spec.accessModes,
-          ele.spec.volumeClaimTemplates[0].spec.resources.requests.storage
-        )
-        const newKStatefulSet = new kStatefulSet(
-          ele.metadata.namespace,
-          ele.kind,
-          ele.metadata.name,
-          ele.spec.replicas,
-          ele.spec.serviceName,
-          newStatefulContainer,
-          newVolumeClaimTemplates
-        )
-        kObjArray.push(newKStatefulSet)
-      }
-      else if(ele.kind === 'Service'){
-        // console.log(ele)
-        const newkSerivce = new kService(
-          ele.metadata.namespace,
-          ele.metadata.name,
-          ele.kind,
-          ele.spec.ports[0].port,
-          ele.spec.ports[0].targetPort,
-          ele.spec.selector.name ? ele.spec.selector.name : ele.spec.selector['app.kubernetes.io/name'],
-          ele.spec.type,
-        )
-        kObjArray.push(newkSerivce)
-      }
-      else{
-        console.log("ele",ele)
-      }
-    };
-    SetDataProp(kObjArray)
-    
+    const data = dataParser.parseData(relevantData);
+    SetDataProp(data)
   }
 
   return( !nodeViewPage ?
