+++ src/kObjects/__index.ts
@@ -1,3 +1,7 @@
+/*
+* utility script for kObjects 
+*/
+
 //allow for one line importing
 export {default as container} from './container';
 export {env} from './container';
@@ -9,3 +13,8 @@ export {kStatefulSet} from './kStatefulSet';
 export {default as statefulContainer, volumeMount} from './statefulContainer';
 export {default as volumeClaimTemplates} from './volumeClaimTemplates';
 export {default as kService} from './kService';
+
+//allow for object param declarations
+export interface anyObject {
+  [key: string]: any
+}
